requirements.txt
flask
google-genai
gunicorn
os